+++ src/provider/index.ts
@@ -1,6 +1,7 @@
 import { Provider } from './default';
 
 export * from './default';
+export * from './errors';
 export * from './interface';
 
 export const defaultProvider = new Provider();