+++ src/types/account.ts
@@ -2,9 +2,10 @@ import BN from 'bn.js';
 
 import { BlockIdentifier } from '../provider/utils';
 import { BigNumberish } from '../utils/number';
-import { EstimateFeeResponse } from './api';
 
-export interface EstimateFee extends EstimateFeeResponse {
+export interface EstimateFee {
+  amount: BN;
+  unit: string;
   suggestedMaxFee: BN;
 }